+++ NEWS.md
@@ -4,7 +4,24 @@ NEWS
 
 <!-- NEWS.md is generated from NEWS.Rmd. Please edit that file -->
 
-    #> Last Update: 2020-10-16 11:54:57
+    #> Last Update: 2020-10-19 10:24:16
+
+# LakeMonitoR 0.0.1.9002
+
+Released - 2020-10-19
+
+  - Update daily\_depth\_means function.
+  - Add data file, laketemp\_ddm (output of daily\_depth\_means
+    function).
+
+# LakeMonitoR 0.0.1.9001
+
+Released - 2020-10-16
+
+  - Add laketemp data
+  - Add functions.
+      - daily\_depth\_means
+      - lake\_strat
 
 # LakeMonitoR 0.0.1.9001